requirements.txt
# Flask and core web packages
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
python-dotenv==1.0.1

# AI / Data analysis
pandas==2.1.1
numpy==1.26.2
scikit-learn==1.3.1
plotly==5.16.1
matplotlib==3.8.2
seaborn==0.13.2

# Optional: for handling HTTP requests
requests==2.31.0

# Optional: for security / password hashing
bcrypt==4.0.1
gunicorn==21.2.0

# Ethereum / Blockchain
web3==6.11.3
eth-account==0.9.0
py-solc-x==1.1.1
